+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/instances/EntityInstance.java
@@ -34,7 +34,7 @@ public class EntityInstance extends BuiltinClass {
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), this.value);
+        return Objects.hash(this.value);
     }
 
     @Override